chore: fix all 22 clippy warnings
`cargo clippy --all-targets` was reporting 22 warnings across 15 files,
all of them style lints predating this change. Now clean.

Behavior-preserving throughout, with two worth calling out:

- icon.rs used `PCWSTR(1 as *const u16)` for MAKEINTRESOURCE, which trips
  `manual_dangling_ptr`. Clippy's suggested `std::ptr::dangling::<u16>()`
  would be wrong here — it yields the alignment of u16 (2), not the
  resource ID 1. Uses `std::ptr::without_provenance(1)` instead, which is
  the correct idiom for an integer sentinel in pointer position and keeps
  the value at 1.

- settings_panel.rs had ~170 lines of real items after `mod tests`
  (`items_after_test_module`). The test module moved to the end of the
  file; no test bodies changed.

The rest are mechanical: `&mut Vec<T>` to `&mut [T]` on two sort/assign
helpers, `&mut nid` to `&nid` for Shell_NotifyIconW, `Default::default()`
field reassignment folded into struct literals, redundant `..Default()`
on fully-specified structs dropped, `///` to `//!` for two module docs,
`div_ceil`, `is_ascii_lowercase`, `is_ok_and`, a collapsed `if`, a
redundant `let` binding, and index-based test loops turned into iterators.

Verified: clippy clean, 105 tests pass, and `cargo fmt --check` reports
the same 16 pre-existing diffs as master (no new formatting debt).

+++ src/interaction.rs
@@ -72,7 +72,7 @@ pub fn handle_key_down(
     // Use GetAsyncKeyState (physical key state) instead of GetKeyState
     // because the low-level keyboard hook swallows all keystrokes before the
     // message queue processes them, so GetKeyState never sees Ctrl as pressed.
-    let ctrl_held = unsafe { (GetAsyncKeyState(VK_CONTROL.0 as i32) as i16) < 0 };
+    let ctrl_held = unsafe { GetAsyncKeyState(VK_CONTROL.0 as i32) < 0 };
 
     handle_key_down_with_ctrl_state(vk_code, state, windows, tags, direct_switch, ctrl_held)
 }
